requirements.txt
Django==1.7.1
MySQL-python
Pillow
django-crispy-forms==1.4.0
django-contact-form
python-dateutil
